requirements.txt
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
